2017-4-23

+++ script/Sh13_kool_proxy.sh
@@ -1,26 +1,30 @@
 #!/bin/sh
 #copyright by hiboy
 source /etc/storage/script/init.sh
+TAG="AD_BYBY"		  # iptables tag
+koolproxy_enable=`nvram get koolproxy_enable`
+[ -z $koolproxy_enable ] && koolproxy_enable=0 && nvram set koolproxy_enable=0
+if [ "$koolproxy_enable" != "0" ] ; then
 nvramshow=`nvram showall | grep ss | awk '{print gensub(/'"'"'/,"'"'"'\"'"'"'\"'"'"'","g",$0);}'| awk '{print gensub(/=/,"='\''",1,$0)"'\'';";}'` && eval $nvramshow
 nvramshow=`nvram showall | grep adbyby | awk '{print gensub(/'"'"'/,"'"'"'\"'"'"'\"'"'"'","g",$0);}'| awk '{print gensub(/=/,"='\''",1,$0)"'\'';";}'` && eval $nvramshow
 nvramshow=`nvram showall | grep adm | awk '{print gensub(/'"'"'/,"'"'"'\"'"'"'\"'"'"'","g",$0);}'| awk '{print gensub(/=/,"='\''",1,$0)"'\'';";}'` && eval $nvramshow
 nvramshow=`nvram showall | grep koolproxy | awk '{print gensub(/'"'"'/,"'"'"'\"'"'"'\"'"'"'","g",$0);}'| awk '{print gensub(/=/,"='\''",1,$0)"'\'';";}'` && eval $nvramshow
 
-[ -z $koolproxy_enable ] && koolproxy_enable=0 && nvram set koolproxy_enable=0
 [ -z $adbyby_mode_x ] && adbyby_mode_x=0 && nvram set adbyby_mode_x=0
 
-koolproxyfile="http://firmware.koolshare.cn/binary/KoolProxy/mipsel"
+koolproxyfile="https://kprule.com/downloads/mipsel"
+koolproxyfilecdn="https://kprules-test.b0.upaiyun.com/downloads/mipsel"
 koolproxyfile2="$hiboyfile/koolproxy"
 koolproxyfile22="$hiboyfile2/koolproxy"
 koolproxyfile3="$hiboyfile/7620koolproxy.tgz"
 koolproxyfile33="$hiboyfile2/7620koolproxy.tgz"
 
-TAG="AD_BYBY"		  # iptables tag
 FWI="/tmp/firewall.adbyby.pdcn" # firewall include file
 AD_LAN_AC_IP=`nvram get AD_LAN_AC_IP`
 AD_LAN_AC_IP=${AD_LAN_AC_IP:-"0"}
 lan_ipaddr=`nvram get lan_ipaddr`
 [ -z "$ss_DNS_Redirect_IP" ] && ss_DNS_Redirect_IP=$lan_ipaddr
+fi
 #检查 dnsmasq 目录参数
 confdir=`grep conf-dir /etc/storage/dnsmasq/dnsmasq.conf | sed 's/.*\=//g'`
 if [ -z "$confdir" ] ; then 
@@ -30,11 +34,33 @@ fi
 gfwlist="/r.gfwlist.conf"
 gfw_black_list="gfwlist"
 
-if [ ! -z "$(echo $scriptfilepath | grep -v "/tmp/script/" | grep kool_proxy)" ]  && [ ! -s /tmp/script/_kool_proxy ]; then
+if [ ! -z "$(echo $scriptfilepath | grep -v "/tmp/script/" | grep kool_proxy)" ]  && [ ! -s /tmp/script/_kool_proxy ] ; then
 	mkdir -p /tmp/script
 	ln -sf $scriptfilepath /tmp/script/_kool_proxy
 	chmod 777 /tmp/script/_kool_proxy
 fi
+
+koolproxy_mount () {
+
+ss_opt_x=`nvram get ss_opt_x`
+upanPath=""
+[ "$ss_opt_x" = "3" ] && upanPath="`df -m | grep /dev/mmcb | grep "/media" | awk '{print $NF}' | awk 'NR==1' `"
+[ "$ss_opt_x" = "4" ] && upanPath="`df -m | grep "/dev/sd" | grep "/media" | awk '{print $NF}' | awk 'NR==1' `"
+[ -z "$upanPath" ] && [ "$ss_opt_x" = "1" ] && upanPath="`df -m | grep /dev/mmcb | grep "/media" | awk '{print $NF}' | awk 'NR==1' `"
+[ -z "$upanPath" ] && [ "$ss_opt_x" = "1" ] && upanPath="`df -m | grep "/dev/sd" | grep "/media" | awk '{print $NF}' | awk 'NR==1' `"
+echo "$upanPath"
+if [ ! -z "$upanPath" ] ; then 
+	logger -t "【koolproxy】" "已挂载储存设备, 主程序放外置设备存储"
+	mkdir -p $upanPath/7620koolproxy
+	mkdir -p /tmp/7620koolproxy
+	mount --bind $upanPath/7620koolproxy /tmp/7620koolproxy
+else
+	logger -t "【koolproxy】" "未挂载储存设备, 主程序放路由内存存储"
+	mkdir -p /tmp/7620koolproxy
+fi
+export PATH='/tmp/7620koolproxy:/etc/storage/bin:/tmp/script:/etc/storage/script:/opt/usr/sbin:/opt/usr/bin:/opt/sbin:/opt/bin:/usr/local/sbin:/usr/sbin:/usr/bin:/sbin:/bin'
+hash koolproxy 2>/dev/null || rm -rf /tmp/7620koolproxy/*
+}
 
 koolproxy_check () {
 
@@ -48,7 +74,7 @@ else
 	needed_restart=0
 fi
 if [ "$koolproxy_enable" != "1" ] && [ "$needed_restart" = "1" ] ; then
-	[ ! -z "`pidof koolproxy`" ] && logger -t "【ADM】" "停止 koolproxy" && koolproxy_close
+	[ ! -z "`pidof koolproxy`" ] && logger -t "【koolproxy】" "停止 koolproxy" && koolproxy_close
 	{ eval $(ps - w | grep "$scriptname" | grep -v grep | awk '{print "kill "$1";";}'); exit 0; }
 fi
 if [ "$koolproxy_enable" = "1" ] ; then
@@ -217,9 +243,10 @@ if [ -z "`pidof koolproxy`" ] && [ "$koolproxy_enable" = "1" ] && [ ! -f /tmp/cr
 	do
 		modprobe $module
 	done 
-	if [ "$koolproxy_auto" = "1" ] && [ -s /etc_ro/7620koolproxy.tgz ]; then
+	koolproxy_mount
+	if [ -s /etc_ro/7620koolproxy_3_4.tgz ] && [ ! -s "/tmp/7620koolproxy/koolproxy" ] ; then
 		logger -t "【koolproxy】" "使用内置主程序"
-		untar.sh /etc_ro/7620koolproxy.tgz /tmp /tmp/7620koolproxy/data/version
+		untar.sh /etc_ro/7620koolproxy_3_4.tgz /tmp /tmp/7620koolproxy/data/version
 	fi
 	if [ ! -s "/tmp/7620koolproxy/data/version" ] ; then
 		logger -t "【koolproxy】" "开始下载 7620koolproxy.tgz"
@@ -228,11 +255,11 @@ if [ -z "`pidof koolproxy`" ] && [ "$koolproxy_enable" = "1" ] && [ ! -f /tmp/cr
 	fi
 	if [ ! -s "/tmp/7620koolproxy/koolproxy" ] ; then
 		logger -t "【koolproxy】" "开始下载 koolproxy"
-		wgetcurl.sh /tmp/7620koolproxy/koolproxy $koolproxyfile $koolproxyfile2
+		wgetcurl.sh /tmp/7620koolproxy/koolproxy $koolproxyfile $koolproxyfilecdn
 	fi
 	if [ ! -s "/tmp/7620koolproxy/koolproxy" ] ; then
 		logger -t "【koolproxy】" "开始下载 koolproxy"
-		wgetcurl.sh /tmp/7620koolproxy/koolproxy $koolproxyfile22 $koolproxyfile
+		wgetcurl.sh /tmp/7620koolproxy/koolproxy $koolproxyfile2 $koolproxyfile22
 	fi
 	if [ ! -s "/tmp/7620koolproxy/koolproxy" ] ; then
 		logger -t "【koolproxy】" "下载失败, 注意检查端口是否有冲突,程序是否下载完整,10 秒后自动尝试重新启动" && sleep 10 && { nvram set koolproxy_status=00; eval "$scriptfilepath &"; exit 0; }
@@ -243,7 +270,7 @@ if [ -z "`pidof koolproxy`" ] && [ "$koolproxy_enable" = "1" ] && [ ! -f /tmp/cr
 	[ -f /etc/storage/koolproxy/ca.key.pem ] && cp -f /etc/storage/koolproxy/ca.key.pem /tmp/7620koolproxy/data/private/ca.key.pem
 	[ -f /etc/storage/koolproxy/ca.crt ] && cp -f /etc/storage/koolproxy/ca.crt /tmp/7620koolproxy/data/certs/ca.crt
 	cd /tmp/7620koolproxy/data/
-	if [ ! -f /tmp/7620koolproxy/data/private/ca.key.pem ]; then
+	if [ ! -f /tmp/7620koolproxy/data/private/ca.key.pem ] ; then
 		logger -t "【koolproxy】" "检测到首次运行https，开始生成koolproxy证书，用于https过滤！"
 		chmod 777 /tmp/7620koolproxy/data/gen_ca.sh && sh gen_ca.sh
 		# 保存证书
@@ -253,10 +280,12 @@ if [ -z "`pidof koolproxy`" ] && [ "$koolproxy_enable" = "1" ] && [ ! -f /tmp/cr
 	fi
 	touch index.txt
 	echo 1000 > serial
+	mkdir -p /tmp/7620koolproxy/data/rules
 	logger -t "【koolproxy】" "koolproxy证书位于/etc/storage/koolproxy/"
 	# 
 	
 	# 处理第三方自定义规则 /tmp/rule_DOMAIN.txt
+	/etc/storage/ad_config_script.sh
 	adbyby_adblocks=`nvram get adbyby_adblocks`
 	rm -f /tmp/7620koolproxy/data/user.bin
 	rm -f /tmp/7620koolproxy/data/user.txt
@@ -277,14 +306,32 @@ if [ -z "`pidof koolproxy`" ] && [ "$koolproxy_enable" = "1" ] && [ ! -f /tmp/cr
 		fi
 		done < /tmp/rule_DOMAIN.txt
 	fi
+	
+	# 处理 koolproxy加载规则列表 /etc/storage/koolproxy_rules_list.sh
+	rm -f /tmp/7620koolproxy/user_store.txt
+	while read line
+	do
+	c_line=`echo $line |grep -v "#" |grep -v '*'`
+	if [ ! -z "$c_line" ] ; then
+		logger -t "【koolproxy】" "下载规则:$line"
+		wgetcurl.sh /tmp/7620koolproxy/user2.txt $line
+		grep -v '^!' /tmp/7620koolproxy/user2.txt | sort -u | grep -v "^$" >> /tmp/7620koolproxy/user_store.txt
+		rm -f /tmp/7620koolproxy/user2.txt
+	fi
+	done < /etc/storage/koolproxy_rules_list.sh
+	
 	# 合并规则
 	cat /etc/storage/koolproxy_rules_script.sh | grep -v '^!' | grep -v "^$" > /tmp/7620koolproxy/user.txt
 	grep -v '^!' /tmp/7620koolproxy/user3adblocks.txt | grep -v "^$" >> /tmp/7620koolproxy/user.txt
+	grep -v '^!' /tmp/7620koolproxy/user_store.txt | grep -v "^$" >> /tmp/7620koolproxy/user.txt
 	ln -sf /tmp/7620koolproxy/user.txt /tmp/7620koolproxy/data/user.txt
+	ln -sf /tmp/7620koolproxy/user.txt /tmp/7620koolproxy/data/rules/user.txt
 	cd /tmp/7620koolproxy/
+	rm -f /tmp/7620koolproxy/user2.txt /tmp/7620koolproxy/user3adblocks.txt /tmp/7620koolproxy/user_store.txt
+	
 	# 更新规则
-	hash daydayup 2>/dev/null && update_kp_rules_daydayup
-	hash daydayup 2>/dev/null || update_kp_rules
+	#hash daydayup 2>/dev/null && update_kp_rules_daydayup
+	#hash daydayup 2>/dev/null || update_kp_rules
 	logger -t "【koolproxy】" "启动 koolproxy 程序"
 	chmod 777 /tmp/7620koolproxy/koolproxy
 	export PATH='/opt/usr/sbin:/opt/usr/bin:/opt/sbin:/opt/bin:/usr/local/sbin:/usr/sbin:/usr/bin:/sbin:/bin'
@@ -293,21 +340,27 @@ if [ -z "`pidof koolproxy`" ] && [ "$koolproxy_enable" = "1" ] && [ ! -f /tmp/cr
 	sleep 1
 	/tmp/7620koolproxy/koolproxy -d >/dev/null 2>&1 &
 	rm -f /tmp/adbyby_host.conf
-	hash krdl 2>/dev/null && krdl_ipset
 	if [ "$adbyby_adblocks" = "1" ] ; then
 		logger -t "【koolproxy】" "加载 第三方自定义 规则, 等候10秒"
 		sleep 10
 	else
 		sleep 5
 	fi
+	hash krdl 2>/dev/null && krdl_ipset
 fi
+if [ -s /tmp/7620koolproxy/data/rules/koolproxy.txt ] ; then
+nvram set koolproxy_rules_date_local="`sed -n '1,10p' /tmp/7620koolproxy/data/rules/koolproxy.txt | grep "$(sed -n '1,10p' /tmp/7620koolproxy/data/rules/koolproxy.txt | grep -Eo '[0-9]+-[0-9]+-[0-9]+ [0-9]+:[0-9]+' | sed -n '1p')" | sed 's/[x!]//g' | sed -r 's/-{2,}//g' | sed -r 's/\ {2}//g' | sed -r 's/\ {2}//g'`"
+nvram set koolproxy_rules_nu_local="`cat /tmp/7620koolproxy/data/rules/koolproxy.txt | grep -v ! | wc -l`"
+nvram set koolproxy_video_date_local="`sed -n '1,10p' /tmp/7620koolproxy/data/rules/koolproxy.txt | grep "$(sed -n '1,10p' /tmp/7620koolproxy/data/rules/koolproxy.txt | grep -Eo '[0-9]+-[0-9]+-[0-9]+ [0-9]+:[0-9]+' | sed -n '2p')" | sed 's/[x!]//g' | sed -r 's/-{2,}//g' | sed -r 's/\ {2}//g' | sed -r 's/\ {2}//g'`"
+fi
+
 [ ! -z "`pidof koolproxy`" ] && logger -t "【koolproxy】" "启动成功"
 [ -z "`pidof koolproxy`" ] && logger -t "【koolproxy】" "启动失败, 注意检查端口是否有冲突,程序是否下载完整,10 秒后自动尝试重新启动" && sleep 10 && { nvram set koolproxy_status=00; eval "$scriptfilepath &"; exit 0; }
 koolproxy_add_rules
 rm -f /tmp/7620koolproxy.tgz /tmp/cron_adb.lock
 /etc/storage/ez_buttons_script.sh 3 & #更新按钮状态
 logger -t "【koolproxy】" "守护进程启动"
-koolproxy_cron_job
+#koolproxy_cron_job
 eval "$scriptfilepath keep &"
 }
 
@@ -316,19 +369,12 @@ koolproxy_rules_list () {
 if [ ! -s /etc/storage/koolproxy_rules_list.sh ] ; then
 cat > "/etc/storage/koolproxy_rules_list.sh" <<-\KPR
 # 【添加为#注释行，会自动忽略】
-# 【添加为*停用和删除的规则文件】
-# 静态规则：
-http://koolshare.b0.upaiyun.com/rules/koolproxy.txt
-# 视频规则：
-http://koolshare.b0.upaiyun.com/rules/1.dat
-# 0day规则：
-http://koolshare.b0.upaiyun.com/rules/add_rules.txt
 # 爪机浏览器规则：
-https://coding.net/u/Carseason/p/koolproxy/git/raw/master/koolproxy_wp.txt
+http://koolshare.b0.upaiyun.com/wp/koolproxy_wp.txt
 # 拯救室友规则：
-*http://koolshare.b0.upaiyun.com/rules/filter_game.txt
+#http://koolshare.b0.upaiyun.com/rules/filter_game.txt
 # 【adbyby规则翻译的koolproxy兼容规则 by <dsyo2008> http://koolshare.cn/thread-83553-1-1.html】
-*https://raw.githubusercontent.com/dsyo2008/lazy_for_koolproxy/master/lazy_kp.txt
+#https://raw.githubusercontent.com/dsyo2008/lazy_for_koolproxy/master/lazy_kp.txt
 
 KPR
 fi
@@ -401,24 +447,34 @@ done
 
 krdl_ipset () {
 
+I=30
+while [ ! -f /tmp/7620koolproxy/data/rules/koolproxy.txt ]; do
+		I=$(($I - 1))
+		[ $I -lt 0 ] && break
+		sleep 1
+done
+[ ! -f /tmp/7620koolproxy/data/rules/koolproxy.txt ] && return 0
 # Koolproxy Rules to Domain List
 rm -f /tmp/7620koolproxy/domain.txt /tmp/7620koolproxy/domain2.txt /tmp/7620koolproxy/ip.txt
-cd /tmp/7620koolproxy/data
-while read line
-do
-c_line=`echo $line |grep -v "#" |grep '*'`
-file_name=${line##*/}
-if [ ! -z $file_name ] && [ ! -z "$c_line" ] ; then
-	[ -f ./$file_name ] && rm -f ./$file_name*
-fi
-c_line=`echo $line |grep -v "#" |grep -v '*'`
-file_name=${line##*/}
-if [ ! -z $file_name ] && [ ! -z "$c_line" ] ; then
-	[ -f ./$file_name ] && krdl ./$file_name
-fi
-done < /etc/storage/koolproxy_rules_list.sh
+cd /tmp/7620koolproxy/data/rules
+# while read line
+# do
+# c_line=`echo $line |grep -v "#" |grep '*'`
+# file_name=${line##*/}
+# if [ ! -z $file_name ] && [ ! -z "$c_line" ] ; then
+	# [ -f ./$file_name ] && rm -f ./$file_name*
+# fi
+# c_line=`echo $line |grep -v "#" |grep -v '*'`
+# file_name=${line##*/}
+# if [ ! -z $file_name ] && [ ! -z "$c_line" ] ; then
+	# [ -f ./$file_name ] && krdl ./$file_name
+# fi
+# done < /etc/storage/koolproxy_rules_list.sh
+# krdl ./user.txt
+krdl ./1.dat
+krdl ./koolproxy.txt
 krdl ./user.txt
-eval $(ls| grep txt.http| awk '{print "cat /tmp/7620koolproxy/data/"$1" >> /tmp/7620koolproxy/domain.txt;";}')
+eval $(ls| grep txt.http| awk '{print "cat /tmp/7620koolproxy/data/rules/"$1" >> /tmp/7620koolproxy/domain.txt;";}')
 # 提取IP
 cat  /tmp/7620koolproxy/domain.txt /tmp/7620koolproxy/koolproxy_blockip.txt | grep -Eo '^[0-9\.]*$' | sort -u | grep -v "^$" > /tmp/7620koolproxy/ip.txt
 cat /tmp/7620koolproxy/ip.txt /tmp/7620koolproxy/koolproxy_blockip.txt | sort -u > /tmp/7620koolproxy/koolproxy_blockip.txt
@@ -427,7 +483,7 @@ cat  /tmp/7620koolproxy/domain.txt | grep  -Ev '^[0-9\.]*$' | sort -u > /tmp/762
 sed -e "s/^/ipset=\/\./" -e "s/$/\/black_koolproxy/" -i /tmp/7620koolproxy/domain2.txt
 cat /tmp/7620koolproxy/domain2.txt /tmp/7620koolproxy/data/koolproxy_ipset.conf | sort -u > /tmp/adbyby_host.conf
 # 删tmp
-rm -f /tmp/7620koolproxy/data/*.txt.http /tmp/7620koolproxy/data/*.txt.https
+rm -f /tmp/7620koolproxy/data/rules/*.txt.http /tmp/7620koolproxy/data/rules/*.txt.https
 rm -f /tmp/7620koolproxy/domain.txt /tmp/7620koolproxy/domain2.txt /tmp/7620koolproxy/ip.txt
 }
 
@@ -777,14 +833,14 @@ koolproxy_cron_job(){
 	koolproxy_update=${koolproxy_update:-"0"}
 	koolproxy_update_hour=${koolproxy_update_hour:-"23"}
 	koolproxy_update_min=${koolproxy_update_min:-"59"}
-	if [ "0" == "$koolproxy_update" ]; then
+	if [ "0" == "$koolproxy_update" ] ; then
 	[ $koolproxy_update_hour -gt 23 ] && koolproxy_update_hour=23 && nvram set koolproxy_update_hour=$koolproxy_update_hour
 	[ $koolproxy_update_hour -lt 0 ] && koolproxy_update_hour=0 && nvram set koolproxy_update_hour=$koolproxy_update_hour
 	[ $koolproxy_update_min -gt 59 ] && koolproxy_update_min=59 && nvram set koolproxy_update_min=$koolproxy_update_min
 	[ $koolproxy_update_min -lt 0 ] && koolproxy_update_min=0 && nvram set koolproxy_update_min=$koolproxy_update_min
 		logger -t "【koolproxy】" "开启规则定时更新，每天"$koolproxy_update_hour"时"$koolproxy_update_min"分，检查在线规则更新..."
 		cru.sh a koolproxy_update "$koolproxy_update_min $koolproxy_update_hour * * * $scriptfilepath update &" &
-	elif [ "1" == "$koolproxy_update" ]; then
+	elif [ "1" == "$koolproxy_update" ] ; then
 	#[ $koolproxy_update_hour -gt 23 ] && koolproxy_update_hour=23 && nvram set koolproxy_update_hour=$koolproxy_update_hour
 	[ $koolproxy_update_hour -lt 0 ] && koolproxy_update_hour=0 && nvram set koolproxy_update_hour=$koolproxy_update_hour
 	[ $koolproxy_update_min -gt 59 ] && koolproxy_update_min=59 && nvram set koolproxy_update_min=$koolproxy_update_min
@@ -821,6 +877,7 @@ C)
 	koolproxy_cp_rules
 	;;
 update)
+	exit 0
 	[ "$koolproxy_enable" != "1" ] && exit 0
 	killall sh_ad_kp_keey_k.sh
 	killall -9 sh_ad_kp_keey_k.sh
@@ -848,6 +905,12 @@ update)
 	logger -t "【koolproxy】" "更新检查:不需更新 $urla "
 	}
 	[ -f /tmp/sh_ad_kp_keey_k.sh ] && /tmp/sh_ad_kp_keey_k.sh &
+	;;
+update_ad)
+	koolproxy_mount
+	koolproxy_close
+	rm -rf /tmp/7620koolproxy/*
+	nvram set koolproxy_status=00 && { eval "$scriptfilepath start &"; exit 0; }
 	;;
 *)
 	koolproxy_check
